+++ components/dropdown/user/src/index.js
@@ -30,14 +30,15 @@ class DropdownUser extends Component {
     })
   }
 
-  _renderLink = ({ text, url, icon: Icon }, index) => {
+  _renderLink = ({ text, url, icon: Icon, notifications }, index) => {
     const Link = this.props.linkFactory
 
     return (
       <li key={index} className='sui-DropdownUserMenu-listItem'>
         <Link href={url} className='sui-DropdownUserMenu-listLink' title={text}>
           <Icon svgClass='sui-DropdownUserMenu-listIcon' />
-          <span>{text}</span>
+          <span className='sui-DropdownUserMenu-listText'>{text}</span>
+          {!!notifications && <span className='sui-DropdownUserMenu-listNotification'>{notifications}</span>}
         </Link>
       </li>
     )
@@ -45,12 +46,12 @@ class DropdownUser extends Component {
 
   render () {
     const { expanded, collapseByTouch } = this.state
-    const { user, menu, expandOnMouseOver } = this.props
+    const { user, menu, expandOnMouseOver, hasNotifications = this.props.menu.some(({ notifications }) => Boolean(notifications)) } = this.props
     const { name, avatar } = user
     const wrapperClassName = cx('sui-DropdownUser', {
-      'is-expanded': expanded
+      'is-expanded': expanded,
+      'has-notifications': hasNotifications
     })
-
     return (
       <div
         className={wrapperClassName}
@@ -65,7 +66,9 @@ class DropdownUser extends Component {
             : this._doNothing
           }
         >
-          <img className='sui-DropdownUser-buttonAvatar' src={avatar} />
+          <div className='sui-DropdownUser-buttonAvatarWrap'>
+            <img className='sui-DropdownUser-buttonAvatar' src={avatar} />
+          </div>
           <span className='sui-DropdownUser-buttonText'>{name}</span>
         </div>
         <div className='sui-DropdownUserMenu-wrap'>
@@ -111,7 +114,12 @@ DropdownUser.propTypes = {
     /**
      * Menu links icon.
      */
-    icon: PropTypes.func.isRequired
+    icon: PropTypes.func.isRequired,
+     /**
+     * Menu links notification.
+     */
+    notifications: PropTypes.number
+
   })).isRequired,
   /**
    * Flag to expand on mouse over event.
@@ -120,10 +128,15 @@ DropdownUser.propTypes = {
   /**
    * Factory for the component that will hold the menu links.
    */
-  linkFactory: PropTypes.func
+  linkFactory: PropTypes.func,
+  /**
+   * Hasnotifications to show a badge notification.
+   */
+  hasNotifications: PropTypes.bool
 }
 
 DropdownUser.defaultProps = {
+  notifications: 0,
   expandOnMouseOver: false,
   linkFactory: ({ href, className, children, title }) =>
     <a href={href} className={className} title={title}>{children}</a>