+++ components/dropdown/user/CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.12.0"></a>
+# 1.12.0 (2017-10-19)
+
+
+### Features
+
+* add new class to dropdown user menu list text ([b4dd8df](https://github.com/SUI-Components/sui-components/commit/b4dd8df))
+* align span badge notification to right ([cb9e147](https://github.com/SUI-Components/sui-components/commit/cb9e147))
+
+
+
+<a name="1.11.0"></a>
+# 1.11.0 (2017-10-17)
+
+
+### Features
+
+* add badge notification in dropdown component ([b136c77](https://github.com/SUI-Components/sui-components/commit/b136c77))
+* add notifications badge in dropdown user component ([95783b7](https://github.com/SUI-Components/sui-components/commit/95783b7))
+* add notifications badge in dropdown user component (fix PR) ([fda47a8](https://github.com/SUI-Components/sui-components/commit/fda47a8))
+* add notifications badge in dropdown user component (fix PR) ([c8071c3](https://github.com/SUI-Components/sui-components/commit/c8071c3))
+
+
+
+<a name="1.10.0"></a>
+# 1.10.0 (2017-10-09)
+
+
+### Features
+
+* switch to s-ui/component-dependencies and sui-theme@8 ([e9cbb94](https://github.com/SUI-Components/sui-components/commit/e9cbb94))
+
+
+
 <a name="1.9.0"></a>
 # 1.9.0 (2017-06-14)