+++ components/form/checkboxList/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.3.0"></a>
+# 1.3.0 (2017-10-09)
+
+
+### Features
+
+* switch to s-ui/component-dependencies and sui-theme@8 ([ac075b4](https://github.com/SUI-Components/sui-components/commit/ac075b4))
+
+
+
 <a name="1.2.0"></a>
 # 1.2.0 (2017-04-10)