+++ components/form/rangeDatepicker/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.3.0"></a>
+# 1.3.0 (2017-10-09)
+
+
+### Features
+
+* switch to s-ui/component-dependencies and sui-theme@8 ([43f2bff](https://github.com/SUI-Components/sui-components/commit/43f2bff))
+
+
+
 <a name="1.2.0"></a>
 # 1.2.0 (2017-09-28)
 
@@ -9,6 +19,21 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 * update path external css of the datepicker ([baa123d](https://github.com/SUI-Components/sui-components/commit/baa123d))
+
+
+
+<a name="1.1.0"></a>
+# 1.1.0 (2017-09-13)
+
+
+### Bug Fixes
+
+* update handleChange event and update scss variables ([7020a3a](https://github.com/SUI-Components/sui-components/commit/7020a3a))
+
+
+### Features
+
+* new component Range datepicker ([412d8f2](https://github.com/SUI-Components/sui-components/commit/412d8f2))