+++ components/image/lazyLoad/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.5.0"></a>
+# 1.5.0 (2017-10-09)
+
+
+### Features
+
+* switch to s-ui/component-dependencies and sui-theme@8 ([a9df35c](https://github.com/SUI-Components/sui-components/commit/a9df35c))
+
+
+
 <a name="1.4.0"></a>
 # 1.4.0 (2017-03-17)