+++ components/topbar/composable/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.3.0"></a>
+# 1.3.0 (2017-10-09)
+
+
+### Features
+
+* switch to s-ui/component-dependencies and sui-theme@8 ([cc80ad6](https://github.com/SUI-Components/sui-components/commit/cc80ad6))
+
+
+
 <a name="1.2.0"></a>
 # 1.2.0 (2017-10-05)
 
@@ -9,6 +19,16 @@ All notable changes to this project will be documented in this file.
 ### Bug Fixes
 
 * rollback className name to keep compatibility ([cfc17ef](https://github.com/SUI-Components/sui-components/commit/cfc17ef))
+
+
+
+<a name="1.1.0"></a>
+# 1.1.0 (2017-10-05)
+
+
+### Features
+
+* move sui-topbar to sui-components ([9829c33](https://github.com/SUI-Components/sui-components/commit/9829c33))